% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{create_lambda_bhq}
\alias{create_lambda_bhq}
\title{Create BHq sequence}
\usage{
create_lambda_bhq(p, fdr)
}
\arguments{
\item{p}{Dimenaion}

\item{fdr}{Expected False Discovery Rate}
}
\description{
Create BHq sequence
}
\examples{
create_lambda_bhq(50, 0.1)
}
