% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{data.generation}
\alias{data.generation}
\title{Data generation}
\usage{
data.generation(n = 100, p = 100, nspr = 10, p.miss = 0.1,
  mu = rep(0, p), Sigma = toeplitz(0^(0:(p - 1))), signallevel = 3,
  mec = "MCAR", sigma = 1, scale = TRUE)
}
\arguments{
\item{n}{Observation number}

\item{p}{Dimenaion}

\item{nspr}{Non-sparsity}

\item{p.miss}{Percentage of missingness}

\item{mu}{Mean of covariates}

\item{Sigma}{Covariance of covariates}

\item{signallevel}{Signal magnitude}

\item{mec}{Missing mechanism}

\item{sigma}{noise variance}

\item{scale}{If TRUE, perform scaling after generating dataset.}
}
\description{
Generate dataset with missing values
}
\examples{
data.list = data.generation()
X = data.list$X.obs
y = data.list$y
}
