% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{slope_admm}
\alias{slope_admm}
\title{Optimization solver}
\usage{
slope_admm(A, b, z, u, lambda_seq, rho, max_iter = 500,
  tol_infeas = 1e-08, verbose = FALSE)
}
\arguments{
\item{A}{Weighted design matrix}

\item{b}{Response}

\item{z}{Initailization}

\item{u}{Initailization}

\item{lambda_seq}{Penalization coefficient}

\item{rho}{Step-size}

\item{max_iter}{Maximum iteration number}

\item{tol_infeas}{Tolerance}

\item{verbose}{If TRUE, print results in each iteration}
}
\description{
Solve optimation with proximal gradient descent
}
